Fix type mismatch warning for option_contrast
- Change extern declaration from 'byte' to 'int' to match saved_data.cpp
- Resolves LTO type mismatch warning during compilation
- Maintains consistency across compilation units

+++ src/signal_meter.cpp
@@ -2,7 +2,7 @@
 
 #ifndef NATIVE_BUILD
 extern PololuLedStrip<12> ledStrip;  // Defined in main.cpp
-extern byte option_contrast;        // Defined in main.cpp
+extern int option_contrast;         // Defined in main.cpp (matches saved_data.cpp type)
 #endif
 
 // Color definitions: Green → Yellow → Red (like analog signal meters)
